Added isDisabled prop to BlockList.Item to disable the Item (#1000)
Signed-off-by: Son Pham <son.pham@optimizely.com>

Before: There's no way to disable an individual item from a BlockList.

Changed: Added an optional isDisabled prop to the BlockList.Item component.

After: If the isDisabled prop is passed in with true, it will only render the prop.children and ignore all other actions. The rendered item is also greyed out and un-clickable. The new prop is optional, so it shouldn't break any existing usage of the BlockList.Item component.

Tests: Added unit-tests and ensure existing tests passed.