Button-Link crossovers (#987)
To standardize how buttons and links are used, we decided to
nest buttons within links and links within buttons. However,
as is, the html would be invalid. So, whenever a Button is
rendered without an onClick, it is converted to a div. When a
Link is rendered without an href, it is converted to a span.
The styling and all the behaviors should still be the same,
it's just that the underlying dom element should be different.

Breaking Change: Link no longer accepts an onClick prop.