feat(FilterPicker): Add FilterPicker Render Component (#1061)
**UPDATES:**
- Created FilterPicker Render Component, a Render Component uses a combination of existing OUI components
- Created story for FilterPicker
- Full set of unit tests for FilterPicker

**ABOUT:**
**_FilterPicker_** is meant to be a replacement for `optimizely.git`'s `react_components/entity_selector` and `react_components/filter_list`.

This component only aims to abstract commonly used display and filter functions, there should be no limit to how extensible it is due to it's [render props](https://reactjs.org/docs/render-props.html) design. While it may be tempting to prioritize lines of code over component composition, I think this pattern will prove to be much more helpful. FilterPicker is React _Render Component_ that uses [render props](https://reactjs.org/docs/render-props.html) to abstract away some of the implementation complexity while allowing for maximum extensibility. This pattern was used to avoid a "props explosion" that would come with all the variations we have of this pattern. This approach will give control to the implementer and keep this component lean.

I used StoryBook to prototype and clean this up, so in order to try it out do the following in the OUI repo (it’s really easy!):
* `git checkout derek/entity_selector_filter_list`
* `nvm use` (if you haven’t done this already)
* `yarn install`
* `yarn storybook`
* Once the Storybook Webstack starts up, go to the `FilterList` component

Some things to note:
* Using `react-immutable-proptypes`, FilterPicker can be used with Immutable or POJO datasets  
* Since Immutable isn't imported into OUI, a pretty interesting Immutable.List/Map _mock_ was used in the test file. This is the best way I can think to do this, but please eye it for any immutable discrepancies. I'll also manually QA in `optimizely.git` before releasing
* By default, this component filters `selectedEntityIds` and items that don't match the `filterQuery` to create and return `availableEntities` via the render props function  
* If a `customFilterFn` is used, it will be used in place of the `filterQuery` filter. If `selectedEntityIds` should be kept in, simply do not include that list as a component prop  
* Find out more in the propType docstrings or the "Show Info" in Storybook

**SCREENSHOT:**
![image](https://user-images.githubusercontent.com/7128855/48517233-813f4080-e81a-11e8-9ded-ae01b13a7fce.png)