patch(FilterPicker): Update scss for search Input (#1090)
**Summary:**
PR #1061 introduced FilterPicker, but overlooked that the class that controls the hiding of the search Input's bottom border was implemented within the render prop. This still is needed, but more explicitly wrapped around the FilterPicker list now and explained in the docs.

For more context, the pattern and story I made with the `oui-filter-picker` was a little unintuitive IMO since it was only in the story and not mentioned in docs. It was also limiting in the case that multiple Inputs are ever needed. So the updates in my PR will make it more clear that `oui-filter-picker-list` class should only wrap BlockList and Input if no bottom border is needed.