docs(FilterPicker): Update Storybook and docs (#1104)
**Summary:**
While at first I thought the `max-scroll--medium` class should go within the `FilterPicker` component itself, I don’t think it really makes sense with the render props pattern it uses. And I think it would be very awkward to allow it to scroll away from the input, which is the only possible way while respecting the render props. We just have to socialize this and ensure people know about the need for it - hence these docs updates.

**Screenshots:**
Without class example:
![screen shot 2019-01-16 at 6 15 08 pm](https://user-images.githubusercontent.com/7128855/51286965-fa921580-19ba-11e9-81c6-eb92c1ddb692.png)

With class example:
![screen shot 2019-01-16 at 6 14 58 pm](https://user-images.githubusercontent.com/7128855/51286963-f5cd6180-19ba-11e9-97e9-412409b54c73.png)