feat(Dropdown): Allow render props with hide function (#1169)
**Summary:**
The goal of this PR is to fix a bug where the toggle function is always called when the Dropdown children elements are clicked, and also to also optionally allow a hide function to be used via the children using a [render props](https://reactjs.org/docs/render-props.html) function. The component will provide the children with the `handleHideChildren` function if the child has the type of function, otherwise it will render the them directly as a node. Unless we expect react to make nodes functions, this will be both backwards compatible and future safe.

Pull this branch and run `yarn storybook` to play with the "Custom hide function" story.

![image](https://user-images.githubusercontent.com/7128855/59118546-11a7b480-8916-11e9-9d3c-578e9a970a2f.png)

**Test plan:**
- [x] Unit tests
- [x] Manual testing