patch(Table/CodeDiff): Revert Table Styles and Update CodeDiff Styles (#1206)
**Summary:**
OUI version `v44.3.0` introduced a CSS regression for how table styles are applied in the monolith, in a large part due to CSS overrides happening there. This reverts the Component css module to the previous state and also brings in some various monolith styles overrides that were happening (e.g. like for `.CalendarDay`). With that, we should highly prioritize cleaning up the overrides outside of OUI and make OUI the complete source of truth.

_Before:_
![image](https://user-images.githubusercontent.com/7128855/62145754-99c48d80-b2b9-11e9-8f85-054e37f1727b.png)

_After:_
![image](https://user-images.githubusercontent.com/7128855/62145788-ad6ff400-b2b9-11e9-9391-c9f042e8a5f5.png)

**Test Plan:**
- [ ] Manually verify tables and nested tables in Monolith and Program Management