perf (OverlayWrapper)[UXENG-27]: Lazily instantiate tether for performance reasons (#1213)
## Summary

The cost of instantiating a Tether component is quite high. In some performance tests, it appears to be more than half the cost of instantiating a table row component.

This PR updates it to lazily instantiate only when its triggered to show. This was tested during hack week and proved to perform quite well.

## Test Plan
Unit tests updated to test new timing instantiation of `_tether`.