Add numRows to Textarea component (#1215)
Currently there is no way to specify the height of the Textarea, which means all Textareas are fixed to a height of 62px. For the textbox used in the create/edit Recommender dialog (for filters) to be replaced with this OUI component, a different height (or numRows) needs to be supported.