feat(NavBar): Added a new Navigation bar component. (#1212)
Summary:
This PR contains a new component which renders the main left navigation bar. It exposes following components.
1. Navbar (The container in which primary links, secondary links and other components will reside)
2. PrimaryLinks - multiple instances passed as children
3. SecondaryLinks - multiple instances passed as children
4. CurrentUserMenu - One instance passed as a child.

Children will always be rendered in order irrespective of the order they were passed in. Primary links will be rendered first, then secondary links and then current user menu.

The idea behind this refactor is to make main navigation generic enough to be used in projects other than FE experimentation. Each project can use this component in its own customized way by implementing a wrapper that provides the necessary state values and customized components and/or dynamic links. This component is stateless and expects state values to be passed from the wrapper. 

Refer to the example in the story to understand how it works.

For more context and detail please read the design doc
[Refactor Global Nav](https://docs.google.com/document/d/1-JGKxEh9an8d5Ic3Um3A7QpFx2aSSbSYBwAY50PnDWM/edit#)

Test Plan:
Added Unit tests for NavBar, IconLink, AccountSwitcher and CurrentUserMenu components

JIRA Issues:
[FEI-551](https://optimizely.atlassian.net/browse/FEI-551)